requirements.txt
opencv-python==4.8.1.78
scikit-image==0.21.0
scikit-learn==1.3.0
seaborn==0.12.2
pandas==2.0.3
matplotlib==3.7.2
numpy==1.24.3
scipy==1.11.1
streamlit==1.27.0
plotly==5.16.1
Pillow==10.0.0 